fix: render issuse
